requirements.txt
readability-lxml
beautifulsoup4
Markdown
requests
Flask
gunicorn
transformers
accelerate
lxml[html_clean]
numpy
torch